Move accessibility modifiers to provider
