+++ src/Features/CSharp/Portable/Formatting/CSharpAccessibilityModifiersNewDocumentFormattingProvider.cs
@@ -1,0 +1,60 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Composition;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.CodeAnalysis.AddAccessibilityModifiers;
+using Microsoft.CodeAnalysis.CodeStyle;
+using Microsoft.CodeAnalysis.Editing;
+using Microsoft.CodeAnalysis.Host.Mef;
+using Microsoft.CodeAnalysis.LanguageServices;
+using Microsoft.CodeAnalysis.Shared.Extensions;
+
+namespace Microsoft.CodeAnalysis.Formatting
+{
+    [Export(LanguageNames.CSharp, typeof(INewDocumentFormattingProvider)), Shared]
+    internal class CSharpAccessibilityModifiersNewDocumentFormattingProvider : INewDocumentFormattingProvider
+    {
+        [ImportingConstructor]
+        [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
+        public CSharpAccessibilityModifiersNewDocumentFormattingProvider()
+        {
+        }
+
+        public async Task<Document> FormatNewDocumentAsync(Document document, CancellationToken cancellationToken)
+        {
+            var documentOptions = await document.GetOptionsAsync(cancellationToken).ConfigureAwait(true);
+            var accessibilityPreferences = documentOptions.GetOption(CodeStyleOptions2.RequireAccessibilityModifiers, document.Project.Language);
+            if (accessibilityPreferences.Value == AccessibilityModifiersRequired.Never)
+            {
+                return document;
+            }
+
+            var root = await document.GetRequiredSyntaxRootAsync(cancellationToken).ConfigureAwait(true);
+            var semanticModel = await document.GetRequiredSemanticModelAsync(cancellationToken).ConfigureAwait(true);
+            var syntaxFacts = document.GetRequiredLanguageService<ISyntaxFactsService>();
+            var typeDeclarations = root.DescendantNodes().Where(node => syntaxFacts.IsTypeDeclaration(node));
+            var editor = new SyntaxEditor(root, document.Project.Solution.Workspace);
+
+            var service = document.GetRequiredLanguageService<IAddAccessibilityModifiersService>();
+
+            foreach (var declaration in typeDeclarations)
+            {
+                if (!service.ShouldUpdateAccessibilityModifier(syntaxFacts, declaration, accessibilityPreferences.Value, out _))
+                    continue;
+
+                var type = semanticModel.GetDeclaredSymbol(declaration, cancellationToken);
+                if (type == null)
+                    continue;
+
+                AddAccessibilityModifiersHelpers.UpdateDeclaration(editor, type, declaration);
+            }
+
+            return document.WithSyntaxRoot(editor.GetChangedRoot());
+        }
+    }
+}
